requirements.txt
# Core Testing Framework
playwright==1.40.0
pytest==7.4.3
pytest-bdd==6.1.1
pytest-playwright==0.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-rerunfailures==12.0
allure-pytest==2.13.2

# Configuration and Data
python-dotenv==1.0.0
faker==20.1.0
pyyaml==6.0.1

# Code Quality Tools
pylint==3.0.3
flake8==7.0.0
black==24.1.1
isort==5.13.2
mypy==1.8.0

# Security Scanning
bandit==1.7.6
safety==3.0.1

# Code Coverage
pytest-cov==4.1.0
coverage==7.4.0

# Pre-commit Hooks
pre-commit==3.6.0

# Performance Testing
pytest-benchmark==4.0.0

# Accessibility Testing
axe-selenium-python==2.1.6

# Additional Utilities
requests==2.31.0
pillow==10.2.0

